fix(tests): deduplicate wiki_dir fixture + add CI coverage enforcement
- #13: Move shared wiki_dir fixture to conftest.py, remove from 3 test files
- #18: Add --cov=athenaeum --cov-fail-under=75 to CI pytest command

Closes #13, closes #18
