+++ tests/test_librarian.py
@@ -4,7 +4,6 @@ process_one, and the run() pipeline with mocked LLM."""
 from __future__ import annotations
 
 import subprocess
-import textwrap
 from pathlib import Path
 from unittest.mock import MagicMock
 
@@ -15,59 +14,6 @@ from athenaeum.models import RawFile
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
-
-
-@pytest.fixture
-def wiki_dir(tmp_path: Path) -> Path:
-    """Create a minimal wiki directory with sample entity pages."""
-    wiki = tmp_path / "wiki"
-    wiki.mkdir()
-
-    (wiki / "feedback_keychain_auth.md").write_text(textwrap.dedent("""\
-        ---
-        name: Auth tokens must use system keychain
-        description: Never store auth tokens as plaintext env vars.
-        type: feedback
-        ---
-
-        Always use the system keychain for storing auth tokens.
-    """))
-
-    (wiki / "a1b2c3d4-acme-corp.md").write_text(textwrap.dedent("""\
-        ---
-        uid: a1b2c3d4
-        type: company
-        name: Acme Corp
-        aliases:
-          - Acme
-          - Acme Corporation
-        access: confidential
-        tags:
-          - client
-          - fintech
-        created: '2024-03-15'
-        updated: '2024-04-06'
-        ---
-
-        # Acme Corp
-
-        Fintech startup, Series B.
-    """))
-
-    (wiki / "project_knowledge_architecture.md").write_text(textwrap.dedent("""\
-        ---
-        name: Knowledge architecture project
-        description: Unified knowledge system.
-        type: project
-        ---
-
-        The knowledge architecture unifies fragmented memory scopes.
-    """))
-
-    (wiki / "_index.md").write_text("# Index\n")
-    (wiki / "MEMORY.md").write_text("# Memory Index\n")
-
-    return wiki
 
 
 @pytest.fixture
